refactor: move remove slice index function to utils so it can be used in extract and make negative controls
