+++ src/readii_2_roqc/utils/metadata.py
@@ -229,3 +229,17 @@ def make_edges_df(mit_index: pd.DataFrame | Path,
     )
         
     return edges_df
+
+
+def remove_slice_index_from_string(img_size:str):
+    """Split up 3D image size into a string that looks like original_##_##_n to remove the slice value and add the original part to the front."""
+
+    split_up_img_size_vals = img_size.split('_')
+    if len(split_up_img_size_vals) > 1:
+        image_size_str = f'original{"_".join(split_up_img_size_vals[0:2])}_n'
+    elif len(split_up_img_size_vals) == 1:
+        image_size_str = f'original{"_".join([split_up_img_size_vals[0], split_up_img_size_vals[0]])}_n'
+    else:
+        image_size_str = 'original'
+
+    return image_size_str
